+++ test/runtests.jl
@@ -535,6 +535,20 @@ resol = Healpix.Resolution(256)
 include("interp.jl")
 include("xyf.jl")
 
+# Iterator interface
+
+m = Healpix.Map{Int8, Healpix.RingOrder}(1)
+for i in 1:length(m)
+    m[i] = i
+end
+
+for i in m
+    @test m[i] == i
+end
+
+@test m[1:3] == [1, 2, 3]
+@test m[end-2:end] == [10, 11, 12]
+
 # Conformability
 
 @test Healpix.conformables(Healpix.Map{Int16,Healpix.RingOrder}(4),